Dev to main

+++ ms-orders/src/main/resources/db/migration/V4__Create_table_products.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `products_tb` (
-  `id` INT AUTO_INCREMENT PRIMARY KEY,
+  `id` BIGINT AUTO_INCREMENT PRIMARY KEY,
   `name` VARCHAR(255),
   `description` VARCHAR(255),
   `price` DOUBLE
